backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
chromadb
pillow
python-multipart
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
pydantic[email]